+++ CHANGELOG.rst
@@ -8,12 +8,14 @@ The format is based on `Keep a Changelog <http://keepachangelog.com/>`_.
 Unpublished
 ====================
 
-NOTE: will be 5.4.x -- drop Python 2.7 support
+IMPORTANT: This release drops support for Python 2 and supports only Python
+3.5 and higher
 
 Added
 _____
 * Cloud only: New regions: mx-monterrey-1, eu-frankfurt-2,
-  eu-madrid-2, eu-jovanovac-1, us-dcc-phoenix-4
+  eu-madrid-2, eu-jovanovac-1, eu-dcc-rome-1, eu-dcc-zurich-1, us-dcc-phoenix-4,
+  us-saltlake-2, sa-valparaiso-1
 * Support for new, flexible wire protocol (V4) has been added. The previous protocol
 is still supported for communication with servers that do not yet support V4. The
 version negotation is internal and automatic; however, use of V4 features will fail